requirements.txt
Django==4.2.7
djangorestframework==3.14.0
djoser==2.2.2
djangorestframework-simplejwt==5.3.0
django-cors-headers
gunicorn